(GH-136) Allow configuration of SHA256 entries
After discussion on Twitch stream, it was decided that it should be
possible to configure the SHA256 entries that are added into the
generated release notes.  This includes showing/hiding the entries,
controlling the heading that is added, as well as the format of each
line that is added to the notes.

To allow for this, there was some general re-formatting of the
codebase, to correctly provide the configuration to each command,
as weel as ensuring default configuration is always provided.

Co-Authored-By: Kim J. Nordmo <admiringworm@users.noreply.github.com>