(maint) Spit up methods throwing exceptions and using async
In general, exceptions shouldn't be thrown when running
inside an async method as they get caught inside an
aggregateexception most/all of the time.

This commit extracts the async code to an async internal
method instead of having the main method being
async. Not all methods are possible to do this way.