(#373) Ensure that release is located correctly
There is no method in Octockit to find a release by a name, only by
a tag. GRM, previously, had a method to filter a returned list of
releases to find the one with the given name, however, this was
incorrectly refactored out, and work had to be done to put it back in
place. Found during testing on Twitch stream.