(#377) Add UploadUrl property to release model
This now seems to be a requirement in Octokit, so we need to make sure
to map this property from the Octokit Release to our internal model, so
that we can map it back again, when required.

This may require later refactoring, once we figure out how to handle
multiple providers.