(build) Fix unit tests
As a result of changes to "how" a release is found from GitHub, it is
no longer the case that an exception is thrown, but rather, if a
specific release can't be found, it simply returns a null value. The
unit tests were updated to reflect this change in functionality.