(build) Force installation of older Cake.Git
When running on GitHub Actions, there is an alias using the Cake.Git
GitPush alias.  This seems to be a problem with the 1.1.0 version of
Cake.Git being used by Cake.Issues.Recipe.  Forcing the
installation/usage of an earlier 0.22.0 version, seems to make things
work locally, so let's test that on GitHub Actions.