Continuous Integration Publish: 789b12db4225c233db2cbce7fd02394ae1374b01 (#513) Re-use existing Config from container
When calling the showconfig command, it was re-hydrating the
configuration file from disk, rather than using the existing
configuration from the IoC container.

This commit changes the associated methods to re-use the existing
Config object, rather than creating it again from scratch.  This has
the result of only showing the Warning about a missing Yaml file once,
and not twice.