(maint) Fix YamlDotNet serializer
In the previous version of YamlDotNet, the Newline property of the
TextWriter was used to control whether a comment was included in the
output, however, this no longer seems to be respected in the latest
version of YamlDotNet.  Instead, make use of the WithNewLine method, and
conditionally set this based on when it is called.