(maint) Improve output when exporting releases
When exporting releases, if not providing a specific tagName, the output
was:

Exporting release null

Which isn't very helpful.

This commit changes this output so that if the tagName is included,
include it in the output, if it isn't, assume that all releases are
being exported, and output this instead.