(#146) Extend milestone with an internal number
Looking at the upcoming work for GitLab, it is clear that we need to
extend, at the very least, the internal GRM representation. This is due
to the fact that GitLab has both an "internal" and "public" Id for the
milestone. It seems to be the case that the internal one, typically a
higher number, is used when updating/changing a milestone, whereas the
public one is used when linking to a milestone within the project. We
need to use both of these, so we need to capture both of them.  For
something like the GitHub provider, which doesn't make this distinction,
we will simply map the single Id to both the internal and public
properties.