(#146) Remove hard-coded values in create template
Within the scriban templates used for creating release notes, there is a
hard-coded query string portion, which is used when linking to
milestones from the generated release notes.  When there was a single
VCS provider, i.e. GitHub, this was fine.  Now that we are introducing
another provider, we need to make this configurable.

This commit introduces a new method, GetMilestoneUrlQueryString(), which
each VCS provider will need to implement, to return the correct value.
In addition, the scriban template has been updated to make use of this,
as well as the model that is passed into the generation of the template.