(#146) Add GitLab VCS Provider
This commit introduces a new GitLabProvider class, and implements the
necessary methods to allow GRM to create release notes on GitLab. This
is made possible by using the NGitLab library. A new option has been
added to the base command, which allows settings of the --provider at
the command line.  The default value for this option is GitHub, so
everything continues to work as it is expected to.