(#146) Prevent usage of unsupported functionality
There are some pieces of functionality that currently either don't make
sense to support in GitLab.  For example, the label command.  This
command was originally created to plug a hole in missing functionality
on GitHub, however the label management in GitLab is far better, and the
crude creation of labels from GRM doesn't make sense.  Therefore, any
attempt to use this command for GitLab returns an error code.

Similarly, the addassets command is not supported. This is mainly due to
the fact that assets work very differently in GitLab and we need to
figure out how to do what is required via the NGitLab library, as I
don't think the support currently exists in there.

Finally, the updating of comments when closing a milestone also isn't
supported, since this also doesn't seem to be supported in NGitLab, but
I will be following up with the team over there to make sure.