(#146) Add breaking API changes
In order to fully support GitLab, we need to introduce come breaking
changes to the GitReleaseManager API.  GitLab has the concept of both a
public and an internal number for things like a milestone. Due to this,
the decision has been made to not assume the property that is required
for a method.  For example, when deleting a release, let's not assume
that we want/need the number of the release, instead, let's pass the
entire release to the method, and then internally, the implementation of
the method can decide what property it needs to use to perform the
operation. This approach as been implemented across all of the API
surface, so we now pass in the entire Release, or Label, or Issue. This
has meant quite a few changes across things like the tests to
accommodate this change.