(#146) Introduce the concept of IsPullRequest
Within GitHub, an issue and a pull request are essentially the same
thing, at least in terms of how GRM uses them.  However, in GitLab, they
are uniquely different, and have their own unique numbers, where as in
GitHub an issue or pull request just get the next number for the
repository. So that we don't have to change too many things in GRM,
let's introduce the concept of an IsPullRequest property to the issue
model, that way the VcsProvider can provide this information if
required, and then use the information to do the right thing when it
comes to fetching and updating the information, but internally, GRM will
just have a collection of issues, which it then acts on when required.