﻿(#89) Include contributors in release notes
This commit adds the ability to query for, and add, information about
the contributors for linked issues and PR's into the generated release
notes.

This is made possible via a new `include-contributors` option in the
create section of the GitReleaseManager.yaml file.  This is false by
default.

In addition, a new scriban template has been created, so allow complete
segregation between release notes that have contributors, and those that
don't. This was done mainly to allow better maintainability going
forward, and to reduce the complexity of the default template.

This has been implemented for both GitHug and GitLab.  For GitHub, it
was necessary to use GraphQL to get the necessary information, where as
with GitLab, the required information could be returned via the REST API.