(#574) Exclude all issues with label
This updates the handling on how we are excluding labels when generating
release notes.
This allows issues to be assigned with a normal label, but still be
excluded if a different label is listed in the exclusion list.

Previously you would need to add and remove labels between drafting a
release notes.