(#439) Add option to disable warning to stderr
This updates the handling of how we output log messages to the console
to instead of blindly outputting warning messages to stderr all the
time, it will instead check if an environment variable called `CI` is
available and set to true, or the user themself have specified the
`--ci` argument.

When CI system is detected, any warnings will instead be outputted to
the normal stdout along with normal informational messages.