(#476) Add validation of input file path handler
This updates how arguments are parsed when attempting to create a new
release. This disallows explicitly the use of a milestone, and an input
file path in the same call.

This is done as there was confusion on why specifying a input file path
was not using the release notes specified. This should make it clearer
to the caller what the issue is.