Use NUnit's new assertion syntax
This is necessary because I merged the develop branch earlier today where the reference to NUnit's NuGet package was upgraded to version 4.x. If I understand the situation correctly, the 'legacy' assertion syntax was supported until version 3.x and was removed in 4.x. That's why my unit tests with the old syntax worked until I rebased my branch earlier today.