﻿(#495) Allow milestones without issues
This commit introduces a new configuration value in the
GitReleaseManager.yaml file, which is disabled by default.  When
enabled, it will be used to control whether the creation of an "empty"
release is allowed.  This will make use of a new "empty" Scriban
template, which will be introduced in the next commit.