(#633) Update to latest NGitLab package version
This addresses an issue which was seen when trying to close milestones
using the GitLab provider. There would be an error thrown regarding
being able to correctly serialize a number into an int32, since the
number was too big.

An upstream change was needed in NGitLab, so switch to usign to a long.
There was a similar error when using GitHub, and a bump in the version
of Octokit being used was required.

This has been tested against a GitLab repository, and the close command
which was previously failing is now working.