(#495) Make use of new empty Scriban template
When creating a release a new template is used when a milestone does not
have any associated issues AND the developer has indicated they want to
allow milestones without issues.

The decision on which template to use has had to be moved.  It can still
be overridden when required, but the calculated value is based on how
many issues have been found, whether to include contributors, etc.