(build) Don't use Cli version for  dogfooding
The globbing pattern that is used to find a GitReleaseManager exe file
includes exe's within the Cli folder, which is a self contained
application. While this seems to run fine locally on my machine,
running it on the CI build isn't working. So let's stick to only using
the .NET Global Tool version of the exe for dogfooding.  In addition,
let's output the exe that is being used for future reference.