(#690) Add support for .NET 8.0 and 9.0
This commit includes bumps to the TargetFrameworks for each project, as
well as a change to allow the tests to work when more TargetFrameworks
are in play.  This obviously didn't happen before, but now that we are
testing across more frameworks, the tests are a bit more brittle, so I
have changed them to be non parallelizable.